chore: inbox issue approval header
