+++ web/components/inbox/sidebar/inbox-list-item.tsx
@@ -1,4 +1,4 @@
-import { FC } from "react";
+import { FC, useEffect } from "react";
 import { observer } from "mobx-react";
 import Link from "next/link";
 import { useRouter } from "next/router";
@@ -39,42 +39,62 @@ export const InboxIssueListItem: FC<TInboxIssueListItem> = observer((props) => {
 
   if (!issue || !inboxIssueDetail) return <></>;
 
+  useEffect(() => {
+    if (issueId === inboxIssueId) {
+      setTimeout(() => {
+        const issueItemCard = document.getElementById(`inbox-issue-list-item-${issueId}`);
+        if (issueItemCard)
+          issueItemCard.scrollIntoView({
+            behavior: "smooth",
+            block: "center",
+          });
+      }, 200);
+    }
+  }, [issueId, inboxIssueId]);
+
   return (
-    <Link
-      key={`${inboxId}_${issueId}`}
-      href={`/${workspaceSlug}/projects/${projectId}/inbox/${inboxId}?inboxIssueId=${issueId}`}
-    >
-      <div
-        id={issue.id}
-        className={`relative min-h-[5rem] cursor-pointer select-none space-y-3 border-b border-custom-border-200 px-4 py-2 hover:bg-custom-primary/5 ${
-          inboxIssueId === issueId ? "bg-custom-primary/5" : " "
-        } ${inboxIssueDetail.status !== -2 ? "opacity-60" : ""}`}
+    <>
+      <Link
+        id={`inbox-issue-list-item-${issue.id}`}
+        key={`${inboxId}_${issueId}`}
+        href={`/${workspaceSlug}/projects/${projectId}/inbox/${inboxId}?inboxIssueId=${issueId}`}
       >
-        <div className="flex items-center gap-x-2">
-          <p className="flex-shrink-0 text-xs text-custom-text-200">
-            {getProjectById(issue.project_id)?.identifier}-{issue.sequence_id}
-          </p>
-          <h5 className="truncate text-sm">{issue.name}</h5>
-        </div>
-        <div className="flex flex-wrap items-center gap-2">
-          <Tooltip tooltipHeading="Priority" tooltipContent={`${issue.priority ?? "None"}`}>
-            <PriorityIcon priority={issue.priority ?? null} className="h-3.5 w-3.5" />
-          </Tooltip>
-          <Tooltip tooltipHeading="Created on" tooltipContent={`${renderFormattedDate(issue.created_at ?? "")}`}>
-            <div className="flex items-center gap-1 rounded border border-custom-border-200 px-2 py-[0.19rem] text-xs text-custom-text-200 shadow-sm">
-              <CalendarDays size={12} strokeWidth={1.5} />
-              <span>{renderFormattedDate(issue.created_at ?? "")}</span>
+        <div
+          className={`relative min-h-[5rem]select-none space-y-3 border-b border-custom-border-200 px-4 py-2 hover:bg-custom-primary/5 cursor-pointer ${
+            inboxIssueId === issueId ? "bg-custom-primary/5" : " "
+          } ${inboxIssueDetail.status !== -2 ? "opacity-60" : ""}`}
+        >
+          <div className="flex items-center justify-between gap-x-2">
+            <div className="flex items-center gap-x-2">
+              <p className="flex-shrink-0 text-xs text-custom-text-200">
+                {getProjectById(issue.project_id)?.identifier}-{issue.sequence_id}
+              </p>
+              <h5 className="truncate text-sm">{issue.name}</h5>
             </div>
-          </Tooltip>
-          <InboxIssueStatus
-            workspaceSlug={workspaceSlug}
-            projectId={projectId}
-            inboxId={inboxId}
-            issueId={issueId}
-            iconSize={14}
-          />
+            <div>
+              <InboxIssueStatus
+                workspaceSlug={workspaceSlug}
+                projectId={projectId}
+                inboxId={inboxId}
+                issueId={issueId}
+                iconSize={14}
+              />
+            </div>
+          </div>
+
+          <div className="flex flex-wrap items-center gap-2">
+            <Tooltip tooltipHeading="Priority" tooltipContent={`${issue.priority ?? "None"}`}>
+              <PriorityIcon priority={issue.priority ?? null} className="h-3.5 w-3.5" />
+            </Tooltip>
+            <Tooltip tooltipHeading="Created on" tooltipContent={`${renderFormattedDate(issue.created_at ?? "")}`}>
+              <div className="flex items-center gap-1 rounded border border-custom-border-200 px-2 py-[0.19rem] text-xs text-custom-text-200 shadow-sm">
+                <CalendarDays size={12} strokeWidth={1.5} />
+                <span>{renderFormattedDate(issue.created_at ?? "")}</span>
+              </div>
+            </Tooltip>
+          </div>
         </div>
-      </div>
-    </Link>
+      </Link>
+    </>
   );
 });
