+++ web/components/inbox/sidebar/inbox-list.tsx
@@ -18,10 +18,10 @@ export const InboxIssueList: FC<TInboxIssueList> = observer((props) => {
 
   if (!inboxIssueIds) return <></>;
   return (
-    <>
+    <div className="overflow-y-auto w-full h-full">
       {inboxIssueIds.map((issueId) => (
         <InboxIssueListItem workspaceSlug={workspaceSlug} projectId={projectId} inboxId={inboxId} issueId={issueId} />
       ))}
-    </>
+    </div>
   );
 });
